+++ metadata/agent-types/.agent-types.json
@@ -13,7 +13,7 @@
       "ID": "F7926101-5099-4FA5-836A-479D9707C818"
     },
     "sync": {
-      "lastModified": "2025-06-15T20:22:02.320Z",
+      "lastModified": "2025-06-16T18:51:30.409Z",
       "checksum": "1fdf686c688c48843bfb8ed8a6991a8d9e0ebd7c41b7aea3361398c8a01b6674"
     }
   }
