+++ metadata/prompts/.prompts.json
@@ -25,16 +25,31 @@
         {
           "fields": {
             "PromptID": "@parent:ID",
+            "ModelID": "@lookup:AI Models.Name=Qwen 3 32B",
+            "VendorID": "@lookup:MJ: AI Vendors.Name=Groq",
+            "Priority": 1
+          },
+          "primaryKey": {
+            "ID": "354C0AC9-9196-4E9E-A6DA-7E0CEBE1EA92"
+          },
+          "sync": {
+            "lastModified": "2025-06-16T18:51:29.939Z",
+            "checksum": "11b69338b8e7ad6ad5dab8bdb2505ba6c310729dcf56faf5481cb26cad3fd18e"
+          }
+        },
+        {
+          "fields": {
+            "PromptID": "@parent:ID",
             "ModelID": "@lookup:AI Models.Name=GPT 4.1",
             "VendorID": "@lookup:MJ: AI Vendors.Name=OpenAI",
-            "Priority": 1
+            "Priority": 2
           },
           "primaryKey": {
             "ID": "0F7988B5-668A-4B53-A071-DFD0CEF7B762"
           },
           "sync": {
-            "lastModified": "2025-06-15T20:22:01.761Z",
-            "checksum": "6b6154f0c49d785c4e5b90e56012aa3cd00ff1a13d77a526dd3f5945ac7e29fd"
+            "lastModified": "2025-06-16T18:51:29.971Z",
+            "checksum": "7d40085a3c03c160da5bdf78a7362b6c54adf625799cbf88d327d5a5c3d1870f"
           }
         }
       ]
@@ -43,7 +58,7 @@
       "ID": "FF7D441F-36E1-458A-B548-0FC2208923BE"
     },
     "sync": {
-      "lastModified": "2025-06-15T20:22:01.761Z",
+      "lastModified": "2025-06-16T18:51:29.971Z",
       "checksum": "caeded9a46eae371be8fdb9e9a69a1772cc14ea598c432a1f39d3136586c48f9"
     }
   },
@@ -81,7 +96,7 @@
             "ID": "15157529-4E12-4FE1-A313-979D236EF8FD"
           },
           "sync": {
-            "lastModified": "2025-06-15T20:22:01.794Z",
+            "lastModified": "2025-06-16T18:51:29.990Z",
             "checksum": "6b6154f0c49d785c4e5b90e56012aa3cd00ff1a13d77a526dd3f5945ac7e29fd"
           }
         }
@@ -91,7 +106,7 @@
       "ID": "B4053E4F-165C-4AED-8E14-3A0A25A28EFE"
     },
     "sync": {
-      "lastModified": "2025-06-15T20:22:01.794Z",
+      "lastModified": "2025-06-16T18:51:29.990Z",
       "checksum": "6db8baccfd7c998333ed11ad3286059762c526cb1cc815e8acc3e1097be538d9"
     }
   },
@@ -124,9 +139,23 @@
             "ModelID": "@lookup:AI Models.Name=GPT 4.1",
             "VendorID": "@lookup:MJ: AI Vendors.Name=OpenAI",
             "Priority": 1
+          },
+          "primaryKey": {
+            "ID": "922159C2-21EE-421E-A5A9-13FEA8FDB908"
+          },
+          "sync": {
+            "lastModified": "2025-06-16T18:51:30.078Z",
+            "checksum": "6b6154f0c49d785c4e5b90e56012aa3cd00ff1a13d77a526dd3f5945ac7e29fd"
           }
         }
       ]
+    },
+    "primaryKey": {
+      "ID": "26949743-4535-4DA6-86F4-0AD76250E17A"
+    },
+    "sync": {
+      "lastModified": "2025-06-16T18:51:30.078Z",
+      "checksum": "c8c87cab8bcd378ec69cb32a30d1598b9f4e2227a5ca792e076743e918f60110"
     }
   },
   {
@@ -158,9 +187,23 @@
             "ModelID": "@lookup:AI Models.Name=GPT 4.1",
             "VendorID": "@lookup:MJ: AI Vendors.Name=OpenAI",
             "Priority": 1
+          },
+          "primaryKey": {
+            "ID": "F6581BA4-049E-4FCD-A064-8E6E34A500DD"
+          },
+          "sync": {
+            "lastModified": "2025-06-16T18:51:30.143Z",
+            "checksum": "6b6154f0c49d785c4e5b90e56012aa3cd00ff1a13d77a526dd3f5945ac7e29fd"
           }
         }
       ]
+    },
+    "primaryKey": {
+      "ID": "1B078367-9AE8-4CB8-BC32-CAAD5DE506D5"
+    },
+    "sync": {
+      "lastModified": "2025-06-16T18:51:30.143Z",
+      "checksum": "61b9a0fecc556b8facc8453b80fae42d8d42d241e8713a90fc42ce068a1f92f3"
     }
   },
   {
@@ -192,9 +235,23 @@
             "ModelID": "@lookup:AI Models.Name=GPT 4.1",
             "VendorID": "@lookup:MJ: AI Vendors.Name=OpenAI",
             "Priority": 1
+          },
+          "primaryKey": {
+            "ID": "D363C45F-A8DE-40F1-8CB8-0F737DC0F6F2"
+          },
+          "sync": {
+            "lastModified": "2025-06-16T18:51:30.208Z",
+            "checksum": "6b6154f0c49d785c4e5b90e56012aa3cd00ff1a13d77a526dd3f5945ac7e29fd"
           }
         }
       ]
+    },
+    "primaryKey": {
+      "ID": "D9CF490A-5F1B-4B78-A93C-021707D107A2"
+    },
+    "sync": {
+      "lastModified": "2025-06-16T18:51:30.208Z",
+      "checksum": "0ddccf28f624b212d07eee4711448d59326a841d1a528ebd7490595d9141fa73"
     }
   },
   {
@@ -226,9 +283,23 @@
             "ModelID": "@lookup:AI Models.Name=GPT 4.1",
             "VendorID": "@lookup:MJ: AI Vendors.Name=OpenAI",
             "Priority": 1
+          },
+          "primaryKey": {
+            "ID": "876700E8-BCB5-428F-9647-C8DC4E3A3148"
+          },
+          "sync": {
+            "lastModified": "2025-06-16T18:51:30.271Z",
+            "checksum": "6b6154f0c49d785c4e5b90e56012aa3cd00ff1a13d77a526dd3f5945ac7e29fd"
           }
         }
       ]
+    },
+    "primaryKey": {
+      "ID": "DDBD7727-0E89-47D5-AA24-C581648775A0"
+    },
+    "sync": {
+      "lastModified": "2025-06-16T18:51:30.271Z",
+      "checksum": "c01066d2f0c25e85f786c440163246f75c11a1c0c23c4c52370db9bd13e9d234"
     }
   },
   {
@@ -260,9 +331,23 @@
             "ModelID": "@lookup:AI Models.Name=GPT 4.1",
             "VendorID": "@lookup:MJ: AI Vendors.Name=OpenAI",
             "Priority": 1
+          },
+          "primaryKey": {
+            "ID": "C4E749A0-9674-4FC4-969D-C9C93569033C"
+          },
+          "sync": {
+            "lastModified": "2025-06-16T18:51:30.335Z",
+            "checksum": "6b6154f0c49d785c4e5b90e56012aa3cd00ff1a13d77a526dd3f5945ac7e29fd"
           }
         }
       ]
+    },
+    "primaryKey": {
+      "ID": "D536DD1C-84C0-46B6-8AC7-7B332E2B0273"
+    },
+    "sync": {
+      "lastModified": "2025-06-16T18:51:30.335Z",
+      "checksum": "8fe812588f4d89958763d4aecfb7b86091053f37c067e797c63d2e37b7768999"
     }
   },
   {
@@ -294,9 +379,23 @@
             "ModelID": "@lookup:AI Models.Name=GPT 4.1",
             "VendorID": "@lookup:MJ: AI Vendors.Name=OpenAI",
             "Priority": 1
+          },
+          "primaryKey": {
+            "ID": "8427B90E-D304-43BD-9037-1A7C8228C99A"
+          },
+          "sync": {
+            "lastModified": "2025-06-16T18:51:30.401Z",
+            "checksum": "6b6154f0c49d785c4e5b90e56012aa3cd00ff1a13d77a526dd3f5945ac7e29fd"
           }
         }
       ]
+    },
+    "primaryKey": {
+      "ID": "5182F85B-C86F-4FCB-A38E-90687F8D0CD2"
+    },
+    "sync": {
+      "lastModified": "2025-06-16T18:51:30.401Z",
+      "checksum": "034fc33219f7d292975ae9bfedfc4683821055e565de017d69a3200077b06022"
     }
   }
 ]
