+++ packages/AI/Engine/src/index.ts
@@ -1,1 +1,2 @@
 export * from './AIEngine';
+export * from './agent-types';
