+++ packages/Angular/Explorer/core-entity-forms/package.json
@@ -34,6 +34,7 @@
     "@memberjunction/ng-code-editor": "2.50.0",
     "@memberjunction/ng-timeline": "2.50.0",
     "@memberjunction/ng-join-grid": "2.50.0",
+    "@memberjunction/ai-engine-base": "2.50.0",
     "@progress/kendo-angular-grid": "16.2.0",
     "@progress/kendo-angular-dropdowns": "16.2.0",
     "tslib": "^2.3.0"
